requirements.txt
fastapi==0.79.0
pymongo==4.1.1
python-dotenv==0.20.0
